[REFACT] : 리뷰 수정
