+++ src/main/java/domain/ladder/LadderLine.java
@@ -3,7 +3,6 @@ package domain.ladder;
 import java.util.List;
 
 public record LadderLine(
-        List<Boolean> connections,
-        String line
+        List<Boolean> connections
 ) {
 }
