+++ src/main/java/service/results/ResultsService.java
@@ -5,7 +5,7 @@ import domain.results.Results;
 import java.util.List;
 
 public interface ResultsService {
-    void makeResults(final List<String> results);
+    void saveResults(final List<String> results);
 
     Results getResults();
 }
